+++ CHANGELOG.md
@@ -4,11 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2022-11-23
+## [Unreleased] - 2022-11-25
 
 ### Added
 
 ### Changed
+* Refactored test cases for the LargestCommonSubgraph class (based on RefactorFirst scan).
 
 ### Deprecated